requirements.txt
Pillow>=10.0.0
PyPDF2>=3.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
numpy>=1.22.0,<1.28.0
py7zr>=0.22.0
rarfile>=4.0
typing-extensions>=4.13.2
websocket-client>=1.8.0